intern

Module intern

Interning of single values.

Interning supports two modes: GC and non-GC.

In non-GC mode, you create Interneds, and can create Copy handles to them that can still be upgraded back to Interned (InternedRef) via Interned::as_ref. Generally, letting the InternedRef to outlive the Interned is a soundness bug and can lead to UB. When all Interneds of some value are dropped, the value is freed (newer interns may re-create it, not necessarily in the same place).

In GC mode, you generally operate on InternedRefs. They are Copy and comfortable. To intern a value you call Interned::new_gc, which returns an InternedRef. Having all Interneds of some value be dropped will not immediately free the value. Instead, a mark-and-sweep GC can be initiated, which will free all values which have no live Interneds.

Generally, in GC mode, you operate on InternedRef, but when you need to store some long-term value (e.g. a Salsa query output), you convert it to an Interned. This ensures that an eventual GC will not free it as long as it is alive.

Making mistakes is hard due to GC InternedRef wrappers not implementing salsa::Update, meaning Salsa will ensure you do not store them in queries or Salsa-interneds. However it’s still possible without unsafe code (for example, by storing them in a static), which is why triggering GC is unsafe.

For more information about GC see crate::gc.

Structs

InternStorage

Interned

InternedRef

Traits

Internable

Type Aliases

Guard 🔒

InternMap 🔒