intern

Struct GarbageCollector

pub struct GarbageCollector {
    alive: FxHashSet<usize>,
    storages: Vec<&'static (dyn Storage + Send + Sync)>,
}

Fields
alive: FxHashSet<usize>storages: Vec<&'static (dyn Storage + Send + Sync)>
Implementations

impl GarbageCollector

pub fn add_storage<T: Internable + GcInternedVisit>(&mut self)

pub fn add_slice_storage<T: SliceInternable + GcInternedSliceVisit>(&mut self)

pub unsafe fn collect(self)

Safety

This cannot be called if there are some not-yet-recorded type values.

All relevant storages must have been added; that is, within the full graph of values, the added storages must form a DAG.

GcInternedVisit and GcInternedSliceVisit must mark all values reachable from the node.

pub fn mark_interned_alive<T: Internable>( &mut self, interned: InternedRef<'_, T>, ) -> ControlFlow<()>

pub fn mark_interned_slice_alive<T: SliceInternable>( &mut self, interned: InternedSliceRef<'_, T>, ) -> ControlFlow<()>

fn sweep_storage<T: Hash + Eq + Send + Sync>( &self, storage: &DashMap<T, (), FxBuildHasher>, get_addr: impl Fn(&T) -> usize + Send + Sync, )

fn retain_only_alive<T>( &self, map: &mut RawTable<T>, get_addr: impl FnMut(&T) -> usize, )