intern

Struct InternedSliceRef

#[repr(transparent)]
pub struct InternedSliceRef<'a, T> {
    ptr: NonNull<c_void>,
    _marker: PhantomData<&'a T>,
}

Fields
ptr: NonNull<c_void>

Invariant

There is no ThinArcBorrow unfortunately, so this is basically a ManuallyDrop<ThinArc>, except that canโt be Copy, so we store a raw pointer instead.

_marker: PhantomData<&'a T>
Implementations

impl<'a, T: SliceInternable> InternedSliceRef<'a, T>

fn arc(self) -> ManuallyDrop<ThinArc<T::Header, T::SliceType>>

pub fn to_owned(self) -> InternedSlice<T>

pub fn get( self, ) -> &'a HeaderSlice<HeaderWithLength<<T as SliceInternable>::Header>, [<T as SliceInternable>::SliceType]>

pub unsafe fn decrement_refcount(self)

Safety

You have to make sure the data is not referenced after the refcount reaches zero; beware the interning map also keeps a reference to the value.

pub(crate) fn strong_count(self) -> usize

pub(crate) fn as_raw(self) -> *const c_void

pub fn change_lifetime<'b>(self) -> InternedSliceRef<'b, T>

Available only on GC mode.

Changes the attached lifetime, as in GC mode, the lifetime is more kind of a lint to prevent misuse than actual soundness check.