intern

Module symbol

Attempt at flexible symbol interning, allowing to intern and free strings at runtime while also supporting compile time declaration of symbols that will never be freed.

Modules

symbols

Module defining all known symbols required by the rest of rust-analyzer.

Structs

Symbol

TaggedArcPtr 🔒

A pointer that points to a pointer to a str, it may be backed as a &'static &'static str or Arc<Box<str>> but its size is that of a thin pointer. The active variant is encoded as a tag in the LSB of the alignment niche.

Statics

MAP 🔒

Functions

increase_arc_refcount 🔒

Type Aliases

Map 🔒