Crate intern

Global Arc-based object interning infrastructure.

Eventually this should probably be replaced with salsa-based interning.

Modules

gc 🔒

Garbage collection of interned values.

intern 🔒

Interning of single values.

intern_slice 🔒

Interning of slices, potentially with a header.

sym

Module defining all known symbols required by the rest of rust-analyzer.

symbol 🔒

Attempt at flexible symbol interning, allowing to intern and free strings at runtime while also supporting compile time declaration of symbols that will never be freed.

Structs

GarbageCollector

InternSliceStorage

InternStorage

Interned

InternedRef

InternedSlice

InternedSliceRef

Symbol

Traits

GcInternedSliceVisit

GcInternedVisit

Internable

SliceInternable